+++ src/integrations/scia_integration/scia_load_generators.py
@@ -70,7 +70,11 @@ def get_load_mode_from_params(params: BridgeParams) -> LoadMode:
 
     if berekeningsniveau == "Theoretische wegindeling":
         return LoadMode.THEORETICAL
-    if berekeningsniveau in ["Werkelijke wegindeling", "Werkelijke wegindeling onderliggend wegennet"]:
+    if berekeningsniveau in [
+        "Werkelijke wegindeling",
+        "Werkelijke wegindeling onderliggend wegennet",
+        "Werkelijke wegindeling onderliggend wegennet met bebording",
+    ]:
         return LoadMode.ACTUAL
     # This should never happen with radio button, but fallback for safety
     return LoadMode.THEORETICAL