+++ CHANGELOG.md
@@ -10,7 +10,7 @@
   - Units are consistently applied from data extraction through user interface display
 - **Traffic load cases**: Added tandem loads and udl for real road layout
   - Added functionality dependent on radio button for road layout
-- **Calculation level**: Added the option for calculation level "werkelijke wegindeling onderliggend wegennet, with different load factors for tandem systems and UDL.
+- **Calculation level**: Added the option for calculation level "werkelijke wegindeling onderliggend wegennet" and "werkelijke wegindeling onderliggend wegennet met bebording", with different load factors for tandem systems and UDL.
 
 ### Changed
 - **Refactoring code SCIA load generation**: Refactored code for the loads helper functions